+++ src/pleistocene/assets/scripts/comic-lines.js
@@ -89,12 +89,10 @@
                 drawLine(svg, ...coords);
             }
         }
+        document.body.style.visibility = "visible";
     }
-    const refresh = () => {
-        window.requestAnimationFrame(draw);
-    }
-    window.addEventListener("load", refresh);
-    window.addEventListener("resize", refresh, { useCapture: true });
-    window.addEventListener("scroll", refresh, { useCapture: true });
+    window.addEventListener("load", draw, { useCapture: true });
+    window.addEventListener("resize", draw, { useCapture: true });
+    window.addEventListener("scroll", draw, { useCapture: true });
 })();